refactor shouddown
