+++ csrc/npu/NPUStream.cpp
@@ -17,6 +17,7 @@
 #include "npu/core/interface/AsyncTaskQueueInterface.h"
 #include "npu/core/register/OptionsManager.h"
 #include "npu/acl/include/acl/acl_rt.h"
+#include "npu/adapter/acl_device_adapter.h"
 
 namespace c10_npu {
 namespace {
@@ -264,6 +265,22 @@ NPUStream NPUStream_fromInternals(const LeakyStreamInternals* ptr) {
           c10::Device(c10::DeviceType::PrivateUse1, ptr->device_index),
           NPUStream_getStreamId(ptr)));
 }
+
+aclError SynchronizeUsedDevices() {
+  c10::DeviceIndex cur_device = 0;
+  NPU_CHECK_ERROR(GetDevice(&cur_device));
+  // Synchronize all used devices
+  std::vector<c10::DeviceIndex> device_idx_vec = acl_adapter::GetUsedDevices();
+  for (const auto deviceId : device_idx_vec) {
+    NPU_CHECK_ERROR(SetDevice(deviceId));
+    aclError acl_ret = aclrtSynchronizeDevice();
+    if (acl_ret != ACL_ERROR_NONE) {
+      return acl_ret;
+    }
+  }
+  NPU_CHECK_ERROR(SetDevice(cur_device));
+  return ACL_ERROR_NONE;
+}
 } // namespace
 
 aclrtStream NPUStream::stream() const {
@@ -471,4 +488,20 @@ aclrtStream NPUStream::stream(const bool need_empty) const {
   return stream();
 }
 
+aclError DestroyUsedStreams() {
+  c10::DeviceIndex cur_device = 0;
+  NPU_CHECK_ERROR(GetDevice(&cur_device));
+  // Synchronize all used devices
+  std::vector<c10::DeviceIndex> device_idx_vec = acl_adapter::GetUsedDevices();
+  for (const auto deviceId : device_idx_vec) {
+    NPU_CHECK_ERROR(SetDevice(deviceId));
+    NPUStream stream = getCurrentNPUStream(deviceId);
+    aclError acl_ret = acl::AclrtDestroyStreamForce(stream);
+    if (acl_ret != ACL_ERROR_NONE) {
+      return acl_ret;
+    }
+  }
+  NPU_CHECK_ERROR(SetDevice(cur_device));
+  return ACL_ERROR_NONE;
+}
 } // namespace c10_npu
